# packages/api-schema/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/products": {
      "post": {
        "operationId": "ProductsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProductDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Product created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          }
        },
        "summary": "Create a new product",
        "tags": [
          "products"
        ]
      },
      "get": {
        "operationId": "ProductsController_findAll",
        "parameters": [
          {
            "name": "category",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of products",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProductResponseDto"
                  }
                }
              }
            }
          }
        },
        "summary": "Get all products",
        "tags": [
          "products"
        ]
      }
    },
    "/products/slug/{slug}": {
      "get": {
        "operationId": "ProductsController_findOneBySlug",
        "parameters": [
          {
            "name": "slug",
            "required": true,
            "in": "path",
            "description": "Stable product slug",
            "schema": {}
          }
        ],
        "responses": {
          "200": {
            "description": "Product found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Product not found"
          }
        },
        "summary": "Get a product by slug",
        "tags": [
          "products"
        ]
      }
    },
    "/products/{id}": {
      "get": {
        "operationId": "ProductsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Product UUID",
            "schema": {}
          }
        ],
        "responses": {
          "200": {
            "description": "Product found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Product not found"
          }
        },
        "summary": "Get a product by ID",
        "tags": [
          "products"
        ]
      },
      "put": {
        "operationId": "ProductsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Product UUID",
            "schema": {}
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProductDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Product updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          },
          "404": {
            "description": "Product not found"
          }
        },
        "summary": "Update a product",
        "tags": [
          "products"
        ]
      },
      "delete": {
        "operationId": "ProductsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Product UUID",
            "schema": {}
          }
        ],
        "responses": {
          "200": {
            "description": "Product deleted successfully"
          },
          "404": {
            "description": "Product not found"
          }
        },
        "summary": "Delete a product",
        "tags": [
          "products"
        ]
      }
    },
...
  },
  "info": {
    "title": "KeyMech API",
    "description": "API documentation for KeyMech mechanical keyboard e-commerce platform",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "CreateKeyboardSpecDto": {
        "type": "object",
        "properties": {}
      },
      "CreateSwitchSpecDto": {
        "type": "object",
        "properties": {}
      },
      "CreateKeycapSpecDto": {
        "type": "object",
        "properties": {}
      },
      "CreateVariantDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "sku": {
            "type": "string"
          },
          "extraPrice": {
            "type": "number"
          },
          "images": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "specs": {
            "type": "object"
          }
        },
        "required": [
          "name"
        ]
      },
      "CreateProductDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "price": {
            "type": "number"
          },
          "category": {
            "type": "string",
            "enum": [
              "KEYBOARD",
              "SWITCH",
              "KEYCAP",
              "PLATE",
              "PCB",
              "STABILIZER",
              "ACCESSORY"
            ]
          },
          "status": {
            "type": "string"
          },
          "images": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "gallery": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "soundTests": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "explodedView": {
            "type": "string"
          },
          "technicalSpec": {
            "type": "string"
          },
          "keyboardSpec": {
            "$ref": "#/components/schemas/CreateKeyboardSpecDto"
          },
          "switchSpec": {
            "$ref": "#/components/schemas/CreateSwitchSpecDto"
          },
          "keycapSpec": {
            "$ref": "#/components/schemas/CreateKeycapSpecDto"
          },
          "variants": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CreateVariantDto"
            }
          }
        },
        "required": [
          "name",
          "slug",
          "price",
          "category"
        ]
      },
      "ProductVariantResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "productId": {
            "type": "string"
          },
          "sku": {
            "type": "string",
            "nullable": true
          },
          "name": {
            "type": "string"
          },
          "extraPrice": {
            "type": "number"
          },
          "images": {
            "type": "object",
            "nullable": true
          },
          "specs": {
            "type": "object",
            "nullable": true
          }
        },
        "required": [
          "id",
          "productId",
          "name",
          "extraPrice"
        ]
      },
      "ProductResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "price": {
            "type": "number"
          },
          "category": {
            "type": "string",
            "enum": [
              "KEYBOARD",
              "SWITCH",
              "KEYCAP",
              "PLATE",
              "PCB",
              "STABILIZER",
              "ACCESSORY"
            ]
          },
          "status": {
            "type": "string",
            "nullable": true
          },
          "images": {
            "type": "object",
            "nullable": true
          },
          "gallery": {
            "type": "object",
            "nullable": true
          },
          "soundTests": {
            "type": "object",
            "nullable": true
          },
          "variants": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProductVariantResponseDto"
            }
          },
          "keyboardSpec": {
            "type": "object",
            "nullable": true
          },
          "switchSpec": {
            "type": "object",
            "nullable": true
          },
          "keycapSpec": {
            "type": "object",
            "nullable": true
          }
        },
        "required": [
          "id",
          "name",
          "slug",
          "price",
          "category",
          "variants"
        ]
      },
      "UpdateProductDto": {
        "type": "object",
        "properties": {}
      },
...
    }
  }
}